F-12187 - Avoid widening MAC composite expression

+++ src/wolfcose_mac0.c
@@ -495,7 +495,8 @@ int wc_CoseMac0_Create(const WOLFCOSE_KEY* key, int32_t alg,
 
     /* unprotected headers map (with kid if present) */
     if (ret == WOLFCOSE_SUCCESS) {
-        unprotectedEntries = (size_t)(((kid != NULL) && (kidLen > 0u)) ? 1u : 0u);
+        unprotectedEntries = ((kid != NULL) && (kidLen > 0u)) ?
+                             (size_t)1u : (size_t)0u;
         ret = wc_CBOR_EncodeMapStart(&outCtx, unprotectedEntries);
     }
     if ((ret == WOLFCOSE_SUCCESS) && (kid != NULL) && (kidLen > 0u)) {
